changed serial to bigserial and integer to bigint

+++ migrations/2022-02-02-024130_create_tickets/up.sql
@@ -1,10 +1,10 @@
 CREATE TABLE tickets (
-    id SERIAL PRIMARY KEY,
-    category_id INTEGER REFERENCES categories (id),
+    id BIGSERIAL PRIMARY KEY,
+    category_id BIGINT REFERENCES categories (id),
     subject TEXT NOT NULL,
     description TEXT,
     status TEXT NOT NULL,
     priority TEXT NOT NULL,
     created TIMESTAMP NOT NULL,
-    reporter_id INTEGER REFERENCES users (id) NOT NULL
+    reporter_id BIGINT REFERENCES users (id) NOT NULL
 );
